split files containing multiple classes
