+++ core/API/NoDefaultValue.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Matomo - free/libre analytics platform
+ *
+ * @link https://matomo.org
+ * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ *
+ */
+
+namespace Piwik\API;
+
+/**
+ * To differentiate between "no value" and default value of null
+ *
+ */
+class NoDefaultValue
+{
+}
