+++ core/UpdaterErrorException.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Matomo - free/libre analytics platform
+ *
+ * @link https://matomo.org
+ * @license http://www.gnu.org/licenses/gpl-3.0.html GPL v3 or later
+ *
+ */
+
+namespace Piwik;
+
+/**
+ * Exception thrown by updater if a non-recoverable error occurs
+ *
+ */
+class UpdaterErrorException extends \Exception
+{
+}
